Refactor: decompose HomeScreen & BookScreen into components, integrate services (Stage 2)
