+++ lib/screens/book_screen.dart
@@ -1,9 +1,8 @@
 import 'package:flutter/material.dart';
-import 'package:hive_flutter/hive_flutter.dart';
 import '../models/node.dart';
-import '../models/note.dart';
-import '../widgets/node_tile.dart';
-import 'view_item_screen.dart';
+import 'components/progress_card.dart';
+import 'components/day_note_card.dart';
+import 'components/chapter_tree_view.dart';
 
 class BookScreen extends StatefulWidget {
   final Node node;
@@ -24,102 +23,17 @@ class _BookScreenState extends State<BookScreen> {
   late TextEditingController _nameController;
   bool _isEditingTitle = false;
 
-  // Дневная заметка
-  Note? _dayNote;
-  final TextEditingController _noteController = TextEditingController();
-  bool _isEditingNote = false;
-
   @override
   void initState() {
     super.initState();
     _node = widget.node;
     _nameController = TextEditingController(text: _node.name);
-    _loadDayNote();
-  }
-
-  void _loadDayNote() {
-    if (_node.category == 'planner') {
-      final notesBox = Hive.box<Note>('notes');
-      _dayNote = notesBox.values.firstWhere(
-        (n) => n.linkedNodeId == _node.id,
-        orElse: () {
-          // Автоматически создаём заметку для старых дней, где её ещё нет
-          final newNote = Note(content: '', linkedNodeId: _node.id);
-          notesBox.put(newNote.id, newNote);
-          return newNote;
-        },
-      );
-      _noteController.text = _dayNote!.content;
-    }
-  }
-
-  /// Только сохраняет данные заметки, без setState
-  void _saveDayNote() {
-    if (_dayNote == null) return;
-    final newContent = _noteController.text.trim();
-    if (newContent != _dayNote!.content) {
-      _dayNote!.content = newContent;
-      _dayNote!.updatedAt = DateTime.now();
-      final notesBox = Hive.box<Note>('notes');
-      notesBox.put(_dayNote!.id, _dayNote!);
-    }
   }
 
   @override
   void dispose() {
-    _saveDayNote(); // автосохранение заметки при закрытии экрана
     _nameController.dispose();
-    _noteController.dispose();
     super.dispose();
-  }
-
-  void _toggleExpanded(Node node) =>
-      setState(() => node.isExpanded = !node.isExpanded);
-
-  void _openViewScreen(Node node) async {
-    if (node.children.isNotEmpty) {
-      _toggleExpanded(node);
-      return;
-    }
-    await Navigator.push(
-      context,
-      MaterialPageRoute(
-        builder: (context) => ViewItemScreen(
-          bookId: _node.id,
-          node: node,
-          onNodeUpdated: () {
-            widget.onNodeUpdated();
-            setState(() {});
-          },
-        ),
-      ),
-    );
-  }
-
-  List<Widget> _buildChildren(Node node, int depth) {
-    List<Widget> widgets = [];
-    for (var child in node.children) {
-      widgets.add(
-        NodeTile(
-          node: child,
-          depth: depth,
-          bookId: _node.id,
-          onCheckboxChanged: () {
-            child.toggle();
-            widget.onNodeUpdated();
-            setState(() {});
-          },
-          onTap: () => _openViewScreen(child),
-          onExpandToggle: child.children.isNotEmpty
-              ? () => _toggleExpanded(child)
-              : null,
-        ),
-      );
-      if (child.isExpanded && child.children.isNotEmpty) {
-        widgets.addAll(_buildChildren(child, depth + 1));
-      }
-    }
-    return widgets;
   }
 
   void _saveTitle() {
@@ -137,8 +51,6 @@ class _BookScreenState extends State<BookScreen> {
 
   @override
   Widget build(BuildContext context) {
-    final bool isPlanner = _node.category == 'planner';
-
     return Scaffold(
       appBar: AppBar(
         title: _isEditingTitle
@@ -180,105 +92,16 @@ class _BookScreenState extends State<BookScreen> {
       ),
       body: Column(
         children: [
-          Card(
-            margin: const EdgeInsets.all(16),
-            child: Padding(
-              padding: const EdgeInsets.all(16),
-              child: Column(
-                crossAxisAlignment: CrossAxisAlignment.start,
-                children: [
-                  Row(
-                    mainAxisAlignment: MainAxisAlignment.spaceBetween,
-                    children: [
-                      const Text(
-                        'Прогресс',
-                        style: TextStyle(
-                          fontSize: 18,
-                          fontWeight: FontWeight.bold,
-                        ),
-                      ),
-                      Text(
-                        '${_node.completedLeaves}/${_node.totalLeaves}',
-                        style: const TextStyle(fontSize: 16),
-                      ),
-                    ],
-                  ),
-                  const SizedBox(height: 8),
-                  LinearProgressIndicator(
-                    value: _node.totalLeaves > 0
-                        ? _node.completedLeaves / _node.totalLeaves
-                        : 0,
-                    backgroundColor: Colors.grey[200],
-                    color: Colors.blue,
-                    minHeight: 12,
-                    borderRadius: BorderRadius.circular(6),
-                  ),
-                ],
-              ),
-            ),
+          ProgressCard(node: _node),
+          DayNoteCard(node: _node),
+          ChapterTreeView(
+            node: _node,
+            bookId: _node.id,
+            onNodeUpdated: () {
+              widget.onNodeUpdated();
+              setState(() {});
+            },
           ),
-          if (isPlanner) ...[
-            Card(
-              margin: const EdgeInsets.symmetric(horizontal: 16, vertical: 8),
-              child: Padding(
-                padding: const EdgeInsets.all(12),
-                child: Column(
-                  crossAxisAlignment: CrossAxisAlignment.start,
-                  children: [
-                    Row(
-                      mainAxisAlignment: MainAxisAlignment.spaceBetween,
-                      children: [
-                        const Text(
-                          'Заметка дня',
-                          style: TextStyle(
-                            fontSize: 16,
-                            fontWeight: FontWeight.bold,
-                          ),
-                        ),
-                        IconButton(
-                          icon: Icon(
-                            _isEditingNote ? Icons.check : Icons.edit,
-                            size: 20,
-                          ),
-                          onPressed: () {
-                            if (_isEditingNote) {
-                              _saveDayNote();
-                              setState(() => _isEditingNote = false);
-                            } else {
-                              setState(() => _isEditingNote = true);
-                            }
-                          },
-                        ),
-                      ],
-                    ),
-                    const SizedBox(height: 8),
-                    if (_isEditingNote)
-                      TextField(
-                        controller: _noteController,
-                        maxLines: 5,
-                        autofocus: true,
-                        decoration: const InputDecoration(
-                          hintText: 'Введите заметку...',
-                          border: OutlineInputBorder(),
-                        ),
-                      )
-                    else
-                      Text(
-                        _dayNote?.content.isEmpty == true
-                            ? 'Нет заметки'
-                            : _dayNote!.content,
-                        style: TextStyle(
-                          color: _dayNote?.content.isEmpty == true
-                              ? Colors.grey
-                              : null,
-                        ),
-                      ),
-                  ],
-                ),
-              ),
-            ),
-          ],
-          Expanded(child: ListView(children: _buildChildren(_node, 0))),
         ],
       ),
     );
